Controller: add functions `onActivated`, `onDeactivated` and add the ability to emit events
These changes will help to have more control on controllers
by using events. Since the game doesn't store the current
state (and it shouldn't), they are very helpful for clients
(subscribers) to understand what is going on in the game.

To emit event, use the protected function `Controller::emitEvent`.
Note that controller X can't emit an event of controller Y, and
vice versa. Such an attempt will result in a runtime error.